ID Variable Typings
IDs for HTML elements are exposed on the global scope by default, so it's a nice way to access them just in JS land. Very cool! I haven't seen many people do that in a TS/JSDoc environment, which is kind of unfortunate to me. It works really well for a vanilla setup!

I had to rename the IDs because they have to be 'variable name'-friendly in order to show up, at least that I know of. Maybe I should look into that and test it out, if say it might be something like data attributes, which converts casing for you automatically, when using the JS API.