ID Variable References
Using the IDs references from the elements themselves, rather than selecting them on their own each time.

Looking into this more, it seems like this is now a standard feature in terms of showing ID elements on the global scope as properties, but it sounds to be not recommended. So maybe I will not do this as much for temporary projects, we'll see. It does make things really convenient to set up in terms of finding refs for elements though, and I really do like that.