Type Annotations
This adds the type definitions for the current code, without changing any of the behavior of the code at all. I haven't added nullish checks yet either, as most of the time that can cause you to need to refactor parts of the code, and I only want to do that once everything is fully type-safe. That's a big learning step I think I realized I discovered over time. Fixing the code can "only" be done once everything can be determined at the type level. Of course you don't really need to have it, it's doable without, but it's more likely for something to go under the radar, and to just be rediscovered again later on.